Create README.md
With a duel account on my PC running WOW TBC Bagnon worked fine until one the the accounts was not able to drag the single bag.  The 0ther account was normal.  I tried to mirror the settings of one account to the other,  No change.  I set the settings to the default.  No change.  I screwed around with it until now both accounts will not allow the dragging of the bag.  My dual account on my lap top is normal.  I have set my PC to the lap top settings but still no change.  both have 914-252.zip and version 2.5.2  any help will be greatly appreciated.